LayoutTests/canvas/philip/tests/size.attributes.removed.html

<!DOCTYPE html>
<title>Canvas test: size.attributes.removed</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="120" height="60"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.removed.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 120, "canvas.width", "120");
canvas.removeAttribute('width');
_assertSame(canvas.width, 300, "canvas.width", "300");


});
</script>
