LayoutTests/canvas/philip/tests/size.attributes.removed-expected.txt

< [index] >
size.attributes.removed
Removing content attributes reverts to default size
References: size.missing
Actual output:
Expected output:

Passed
